requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
